requirements.txt
git+https://github.com/zparis99/ECoG-foundation-model.git@v1.0.3#egg=ecog_foundation_model
absl-py==2.1.0
accelerate==0.26.1
aiofiles==22.1.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.32.40
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.3.4
boto3==1.34.35
botocore==1.34.40
braceexpand==0.1.7
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.13.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
docker-pycreds==0.4.0
docutils==0.16
edfio==0.4.3
einops==0.7.0
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.41
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.10.0
hjson==3.1.0
huggingface-hub==0.20.3
humanize==4.9.0
identify==2.5.33
idna==3.6
imageio==2.33.1
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_fileid==0.9.1
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.5.2
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab-nvdashboard==0.9.0
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.6.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
lightning-utilities==0.10.1
lxml==5.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mne==1.6.1
mne-bids==0.14
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
nilearn==0.10.3
ninja==1.11.1.1
nodeenv==1.8.0
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.5.0
pooch==1.8.1
pre-commit==3.6.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyEDFlib==0.1.37
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
sentry-sdk==1.40.0
setproctitle==1.3.3
shellingham==1.5.4
simpervisor==1.0.0
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==1.0.15
tinycss2==1.2.1
tokenizers==0.15.1
torch==2.1.0
torchio==0.19.5
torchmetrics==1.3.0.post0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.0.7
virtualenv==20.25.0
wandb==0.16.2
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.4
widgetsnbextension==4.0.9
wrapt==1.16.0
xformers==0.0.22.post7
xyzservices==2023.10.1
y-py==0.6.2
yarl==1.9.4
ypy-websocket==0.8.4
